refactor: useEmployerRemittanceData 훅 책임별 분리 (SRP)
- 단일 훅(270줄) → 4개 개별 훅 + 조합 훅(100줄)으로 분리
- useFetchWorkplaces: 근무지 목록 페칭 및 선택 상태 관리
- useFetchWorkers: 근로자 목록 페칭 및 선택 상태 관리
- useFetchWorkRecords: 근무 기록 페칭 및 급여 계산
- useMonthNavigation: 월 이동 네비게이션
- useEmployerRemittanceData: 개별 훅 조합 + UI 상태 관리
